Fixed various compiler warnings (mostly float vs int etc).
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/ai@10676 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ trunk/src/karts/controller/fuzzy_data_manager.hpp
@@ -63,10 +63,10 @@ public :
     void        reset();
     
     // -- Setters --
-    void        setPlayerEvaluation(float e) { m_player_eval = e; }
+    void        setPlayerEvaluation(int e) { m_player_eval = e; }
     
     // -- Getters --
-    float       getPlayerEvaluation()         { return m_player_eval; }
+    int         getPlayerEvaluation()         { return m_player_eval; }
     
     const std::vector<std::vector<PathData*>*>* getPathData(
                                                     unsigned int nodeId ) const;
